+++ core/geometry.py
@@ -3,10 +3,7 @@
 import warnings
 
 import geopandas as gpd
-import momepy
-import networkx as nx
 import numpy as np
-import pandas as pd
 import shapely
 from libpysal import graph
 from scipy import spatial
@@ -259,125 +256,3 @@ def snap_to_targets(edgelines, poly, snap_to, secondary_snap_to=None):
                     stacklevel=2,
                 )
     return to_add, to_split
-
-
-def get_components(edgelines, ignore=None):
-    edgelines = np.array(edgelines)
-    start_points = shapely.get_point(edgelines, 0)
-    end_points = shapely.get_point(edgelines, -1)
-    points = shapely.points(
-        np.unique(
-            shapely.get_coordinates(np.concatenate([start_points, end_points])), axis=0
-        )
-    )
-    if ignore is not None:
-        mask = np.isin(points, ignore)
-        points = points[~mask]
-    # query LineString geometry to identify points intersecting 2 geometries
-    inp, res = shapely.STRtree(shapely.boundary(edgelines)).query(
-        points, predicate="intersects"
-    )
-    unique, counts = np.unique(inp, return_counts=True)
-    mask = np.isin(inp, unique[counts == 2])
-    merge_res = res[mask]
-    merge_inp = inp[mask]
-    closed = np.arange(len(edgelines))[shapely.is_closed(edgelines)]
-    mask = np.isin(merge_res, closed) | np.isin(merge_inp, closed)
-    merge_res = merge_res[~mask]
-    merge_inp = merge_inp[~mask]
-    g = nx.Graph(list(zip((merge_inp * -1) - 1, merge_res, strict=True)))
-    components = {
-        i: {v for v in k if v > -1} for i, k in enumerate(nx.connected_components(g))
-    }
-    component_labels = {value: key for key in components for value in components[key]}
-    labels = pd.Series(component_labels, index=range(len(edgelines)))
-
-    max_label = len(edgelines) - 1 if pd.isna(labels.max()) else int(labels.max())
-    filling = pd.Series(range(max_label + 1, max_label + len(edgelines) + 1))
-    labels = labels.fillna(filling)
-
-    return labels.values
-
-
-def weld_edges(edgelines, ignore=None):
-    """lightweight version of remove_false_nodes
-
-    optionally ignore some nodes - do not weld lines
-    """
-    if len(edgelines) < 2:
-        return edgelines
-    labels = get_components(edgelines, ignore=ignore)
-    return (
-        gpd.GeoSeries(edgelines)
-        .groupby(labels)
-        .agg(lambda x: shapely.line_merge(shapely.GeometryCollection(x.values)))
-    ).tolist()
-
-
-def remove_false_nodes(gdf, aggfunc="first", **kwargs):
-    """Reimplementation of momepy.remove_false_nodes that preserves attributes
-
-    Parameters
-    ----------
-    gdf : _type_
-        _description_
-
-    Returns
-    -------
-    _type_
-        _description_
-    """
-    if len(gdf) < 2:
-        return gdf
-
-    if isinstance(gdf, gpd.GeoSeries):
-        gdf = gdf.to_frame("geometry")
-
-    labels = get_components(gdf.geometry)
-
-    # Process non-spatial component
-    data = gdf.drop(labels=gdf.geometry.name, axis=1)
-    aggregated_data = data.groupby(by=labels).agg(aggfunc, **kwargs)
-    aggregated_data.columns = aggregated_data.columns.to_flat_index()
-
-    # Process spatial component
-    def merge_geometries(block):
-        merged_geom = shapely.line_merge(shapely.GeometryCollection(block.values))
-        return merged_geom
-
-    g = gdf.groupby(group_keys=False, by=labels)[gdf.geometry.name].agg(
-        merge_geometries
-    )
-    aggregated_geometry = gpd.GeoDataFrame(g, geometry=gdf.geometry.name, crs=gdf.crs)
-    # Recombine
-    aggregated = aggregated_geometry.join(aggregated_data)
-
-    nodes = momepy.nx_to_gdf(
-        momepy.node_degree(momepy.gdf_to_nx(aggregated[[aggregated.geometry.name]])),
-        lines=False,
-    )
-    loop_mask = aggregated.is_ring
-    loops = aggregated[loop_mask]
-
-    fixed_loops = []
-    fixed_index = []
-    node_ix, loop_ix = loops.sindex.query(nodes.geometry, predicate="intersects")
-    for ix in np.unique(loop_ix):
-        loop_geom = loops.geometry.iloc[ix]
-        target_nodes = nodes.geometry.iloc[node_ix[loop_ix == ix]]
-        if len(target_nodes) == 2:
-            node_coords = shapely.get_coordinates(target_nodes)
-            coords = np.array(loop_geom.coords)
-            new_start = (
-                node_coords[0]
-                if (node_coords[0] != coords[0]).all()
-                else node_coords[1]
-            )
-            new_start_idx = np.where(coords == new_start)[0][0]
-            rolled_coords = np.roll(coords[:-1], -new_start_idx, axis=0)
-            new_sequence = np.append(rolled_coords, rolled_coords[[0]], axis=0)
-            fixed_loops.append(shapely.LineString(new_sequence))
-            fixed_index.append(ix)
-
-    aggregated.loc[loops.index[fixed_index], aggregated.geometry.name] = fixed_loops
-    return aggregated
